hidden layer example partially done. added boost::numeric::ublas to namespace.

+++ random_gen.h
@@ -3,7 +3,7 @@
 class RandomNGenerator {
 
 public:
-	static double returnRandomD(int start, int end) {
+	static double returnRandomD(double start, double end) {
 		std::random_device rd;
 		std::mt19937 gen(rd());
 		std::uniform_real_distribution<> dis(start,end);
